create ajax ok

+++ app/controllers/users_controller.rb
@@ -1,22 +1,18 @@
 class UsersController < ApplicationController
+	skip_before_action :verify_authenticity_token, :only => [:create,:destroy]
 	def index
-		@users=User.page(params[:page]).per(params[:per_page])
+		@users=User.page(params[:page]).per(params[:per_page]||10)
 	end
 	def create
 		@user=User.new(user_params)
-		@create_success=@user.save		
+		@user.save	
 	end
 	def destroy
 		user=User.find_by(params[:id])
-		if user
-			user.destroy
-			@destroy_success=true
-		else
-			@destroy_success=false
-		end		
+		user.destroy if user			
 	end
 	private
 	def user_params
-		params.permit(:name,:email)
+		params.require(:user).permit(:name,:email)
 	end
 end
